backend/auth_service/requirements.txt
# Auth Microservice — Python dependencies
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database (async SQLAlchemy + PostgreSQL driver)
sqlalchemy==2.0.30
asyncpg==0.29.0

# Password hashing
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# JWT
python-jose[cryptography]==3.3.0

# Env / config
python-dotenv==1.0.1

# Request body validation (ships with FastAPI, pinned for clarity)
pydantic[email]==2.7.1
